# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "kaas-ui",
    "description": "A read-only, multi-cluster Kafka UI. Every data route is a GET; there is no mutating endpoint in this document because there is none in the router.",
    "license": {
      "name": "Apache-2.0"
    },
    "version": "0.9.0"
  },
  "paths": {
...
    "/api/environments/{env}/clusters/{id}/groups/{group}/offsets": {
      "get": {
        "tags": [
          "groups"
        ],
        "summary": "`GET /api/clusters/{id}/groups/{group}/offsets`",
        "description": "Committed offsets joined to the log ends, so lag can be classified rather\nthan subtracted. \"No commit yet\", \"empty partition\", \"caught up\" and\n\"behind by n\" are four different answers and rendering them all as `0` is\nhow a lag column becomes something nobody trusts.",
        "operationId": "offsets",
        "parameters": [
          {
            "name": "env",
            "in": "path",
            "description": "Environment id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "Cluster id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "group",
            "in": "path",
            "description": "Group id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Committed offsets and lag",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_GroupOffset"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/environments/{env}/clusters/{id}/topics/{topic}/messages": {
      "get": {
        "tags": [
          "messages"
        ],
        "summary": "`GET /api/clusters/{id}/topics/{topic}/messages`",
        "description": "One bounded page, as a plain JSON resource. The same seven modes as the\nstream and the same rows — a \"load more\" button must not produce rows that\ndiffer in shape from the ones already on screen.",
        "operationId": "page",
        "parameters": [
          {
            "name": "env",
            "in": "path",
            "description": "Environment id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "Cluster id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "topic",
            "in": "path",
            "description": "Topic name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "mode",
            "in": "query",
            "description": "Which window",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SeekMode"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "For fromOffset and toOffset",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "timestamp",
            "in": "query",
            "description": "Epoch millis, for sinceTime and toTime",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "partitions",
            "in": "query",
            "description": "Comma-separated partition list",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "visibility",
            "in": "query",
            "description": "all | committed",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filter",
            "in": "query",
            "description": "Substring match on the value",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Records in this page",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "keyCodec",
            "in": "query",
            "description": "Override how keys are read",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Codec"
            }
          },
          {
            "name": "valueCodec",
            "in": "query",
            "description": "Override how values are read",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Codec"
            }
          },
          {
            "name": "predicate",
            "in": "query",
            "description": "JavaScript expression over the decoded value",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "One page of messages",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessagePage"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/environments/{env}/clusters/{id}/topics/{topic}/messages/tail": {
      "get": {
        "tags": [
          "messages"
        ],
        "summary": "`GET /api/clusters/{id}/topics/{topic}/messages/tail`",
        "description": "`TailSpec::limit` is a per-topic target that kaas-lib spreads across\npartitions with `div_ceil`, so asking for 20 on a 16-partition topic\ngenuinely fetches 32. \"The last n of a topic\" has no single answer across\npartitions, so this layer picks one and says so: **the n most recent by\ntimestamp**, merged across partitions and truncated after the fact.\n`total` reports how many were fetched before truncation.",
        "operationId": "tail",
        "parameters": [
          {
            "name": "env",
            "in": "path",
            "description": "Environment id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "Cluster id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "topic",
            "in": "path",
            "description": "Topic name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Records to return after merging",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "partitions",
            "in": "query",
            "description": "Comma-separated partition list",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "keyCodec",
            "in": "query",
            "description": "Override how keys are read",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Codec"
            }
          },
          {
            "name": "valueCodec",
            "in": "query",
            "description": "Override how values are read",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Codec"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The tail of a topic",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_Message"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/environments/{env}/clusters/{id}/topics/{topic}/offsets": {
      "get": {
        "tags": [
          "topics"
        ],
        "summary": "`GET /api/clusters/{id}/topics/{topic}/offsets`",
        "operationId": "offsets",
        "parameters": [
          {
            "name": "env",
            "in": "path",
            "description": "Environment id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "Cluster id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "topic",
            "in": "path",
            "description": "Topic name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Offset ranges",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_Partition"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Codec": {
        "type": "string",
        "description": "How a payload was read.\n\nThe chip in the message list is this value, and the chip is the **override\ncontrol**, not a label: auto-detection that cannot be corrected is worse\nthan none.",
        "enum": [
          "auto",
          "string",
          "hex",
          "json",
          "avro",
          "protobuf",
          "jsonSchema"
        ]
      },
...
      "SeekMode": {
        "type": "string",
        "description": "How a window of a topic was asked for.",
        "enum": [
          "live",
          "newest",
          "oldest",
          "fromOffset",
          "toOffset",
          "sinceTime",
          "toTime"
        ]
      },
...
    }
  },
...
}